[refactor] : parameterized test
